Setup an index page
I'm not sure why they didn't have this. The site was working because
they have a 404 handler which loads all of the Javascript app and then
redirects to /african-diaspora, so even though it was getting a 404, it
was still displaying the app. This gives the site an index.html page
which works without the 404.

Issue #75733